+++ mtpgsql/src/include/storage/shmem.h
@@ -23,7 +23,7 @@
  * in every process.  Shared memory "pointers" are actually
  * offsets relative to the start of the shared memory region(s).
  */
-typedef unsigned long SHMEM_OFFSET;
+typedef size_t SHMEM_OFFSET;
 
 #define INVALID_OFFSET (-1)
 #define BAD_LOCATION (-1)
